requirements.txt
boto3>=1.34
cartopy>=0.23
matplotlib>=3.8
numpy>=1.26
arm_pyart>=2.1
